Add 'stride' ability in gui (x direction).
With this, we need to store the viewer's position and viewing angles.
Viewing angles are stored in an orthonormal basis (0,0,1) for now.
New basis stuff is in xfrm.c.

A bug is apparent when the view comes thru the side of the bounding box
(x <= 0 || x >= 100). Linear slices of geometry are missed.