Fix a numerical imprecision
which showed a line of missed pixels.
Code's current state shows test case for shit in cli.c (bug not
present).

GUI shows there is still an imprecision issue which was originally seen
when fixing the view plane to the side of the bounding box. The missed
pixels here are much sparser.