Fix another numerical imprecision
As stated last commit, ray misses were seen in the 'perspective' mode
when it was fixed to the side of the geometry's bounding box.

The ray was hitting the bounding box in the Z dimension exactly where
the search space split in the X dimension. It traversed the first search
box, but neglected to continue through to the obvious next one in the
positive X direction. This was so, because the hit_BoundingPlane()
function calculates the entrance point. In this calculation, the Y and Z
components in the ray were not large enough to count as "inside" the
right search box.

The fix assumes the ray passes thru the local bounding box which the
plane partitions. This is true in our use case so the function was
renamed to hit_inner_BoundingPlane(). Intersections computed as stopping
short of the local bounding box are adjusted to be valid.

In short, if a ray enters a box, it cannot exit in the same direction
(for any dimension).