Add OpenCL hello world
The OpenCL code includes a kernel, hello.cl, along with much of the
source tree. Ifdefs are added to make the code compilable to C and
OpenCL where applicable without an ugly refactor.

Works on ATI stream SDK. Problems exist on the Apple OpenCL which could
not be diagnosed today.