Add BitString and partition out sorting
Sorting goes to a new file, order.c/.h.
BitString was added to assist with removal of duplicate elements after an
interpolation, but the algorithm turned out to be invalid. It is used and
a unit test (the first one!) was written for it.

Makefile
* Specify number of dimensions if desired.

bitstring.c/.h
* Add BitString and set/retrieve functions.

kdtree.c
* Move sort algorithms to order.c/.h

order.c/.h
* Grab interval sorting from kdtree.c.

util.c/.h
* Add a ceil() function for  uints.

verif/main.c
* Begin adding a unit test framework. Let us hope it does not grow to be huge.