Stop light from peeking into corners
This was happening:

             wall
               |   eye (can see light!)
         light |  /
             \ | /
              \|/_____floor

And the reason for it was, there was an offset added to the reflection
ray so the reflection would not see the object which the original ray
intersected. Well, the offset was in the reflection direction so in this
case, the "reflection" ray (really a different case, but both were
affected) completely skipped the wall during the offset phase.

1. Fix this behavior by backing up in the original ray (by the offset)
   before casting the reflection ray. This allows us to not accidentally
   skip anything on a reflection, though the ray is slightly incorrect.
   In relation to a pixel size, this is negligible.
  b. raytrace.c  fill_pixel()