Fix long-standing rotation bug
The bug manifested itself when a flying craft acquired traction with
very strong sideways motion such that its new vertical position would be
over 90 degrees away from its current one. The craft's left and right
would flip. In OpenGL mode, this affected only the craft, so turning and
lasers were backwards. In raytrace mode, the entire view flipped with
the craft, which could explain why the bug wasn't noticed until now.

The implementation of orthorotate_PointXfrm() used Gram-Schmidt in a
certain order. A basis vector was perturbed and this method was used to
make the rest of the basis vectors "follow along".
For rotations larger than 90 degrees, elements of basis vectors could
flip, resulting in a matrix with determinant equal to -1.

As a fix, this perturbation-style rotation is re-implemented fully.
Essentially, a plane is formed between the basis vector and its new
vector, rotation is calculated and performed on projections of other
basis vectors onto the plane. The rotated projections are added back
into their respective basis vectors' orthojections (from the plane).
These results are used as the new basis... though the math ends up being
much simpler than this.

1. Remove edge cases from the basis rotation function. It should now
   give a perfectly smooth rotation along the rotation plane.
  b. xfrm.c/h  orthorotate_PointXfrm()

2. Use the new form of the rotation function.
  b. motion.c  apply_thrust()
  c. gui-indep.c  update_view_params()
  c. gui.c  grab_mouse_fn()
  c. motion.c  remove_4d_rotation()
  c. testcase.c  setup_testcase_simple()