Shaders in Windows
To make this run properly, OpenGL drivers need to be installed.
The default Windows wrapper only goes up to version 1.1.
Not sure how to get the right drivers installed under QEMU (or if they
exist), but Wine runs it modulo a couple minor quirks (joystick
wackiness (different mapping?) and a bad message on exit).

1. Windows is dumb and requires function pointers for OpenGL API
   functions past version 1.1. Dynamically look these up.
  a. gui-opengl.c  init_ogl_ui_data()

2. Disable some compile warnings on Windows which I ignore.
  a. util.h
  a. CMakeLists.txt

3. Disable OpenMP on Windows. It's not in my Wine by default.
  c. CMakeLists.txt

4. Move the raw (hand-crafted) Makefile stored in version control
   so CMake doesn't overwrite it.
  c. Makefile.raw

5. Add my Windows path setup script.
  a. paths.bat