Remove cx/
This project is now assumed to be a submodule of outherspace-top.git.
cx/ moves to ../cx/, as another submodule of outherspace-top.git.
Likewise, the data directory is assumed to be in ../outherspace-data/...
though the code which assumes this is still assuming it is in data/.
Just specify files manually instead of using the compiled-in test cases.